requirements.txt
openai==1.97.0
sqlalchemy==2.0.41
pydantic==2.11.7
typing-extensions==4.14.1
python-dotenv==1.0.0
PyMySQL==1.1.1 
dateparser
streamlit
mcp[cli]